third_party/WebKit/LayoutTests/external/wpt/webstorage/storage_builtins-expected.txt

This is a testharness.js-based test.
FAIL Builtins in localStorageassert_array_equals: b property 0, expected function "function key() { [native code] }" but got "key"
FAIL Builtins in sessionStorageassert_array_equals: b property 0, expected function "function key() { [native code] }" but got "key"
Harness: the test ran to completion.
